requirements.txt
flask
openai
requests
python-dotenv
gunicorn
supabase==2.27.2
